+++ src/pages/todo.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react'
-import { useRecoilValue, useSetRecoilState, useRecoilState } from 'recoil'
+import { useRecoilValue, useSetRecoilState } from 'recoil'
+import TodoItem from '../components/todoItem'
 import { todoListState } from '../atoms/todo'
 
 const TodoList = () => {
@@ -50,57 +51,5 @@ let id = 0
 function getId() {
   return id++
 }
-
-const TodoItem = ({ item }) => {
-  const [todoList, setTodoList] = useRecoilState(todoListState)
-  const index = todoList.findIndex((listItem) => listItem === item)
-
-  const editItemText = ({ target: { value } }) => {
-    const newList = replaceItemAtIndex(todoList, index, {
-      ...item,
-      text: value,
-    })
-
-    setTodoList(newList)
-  }
-
-  const toggleItemCompletion = () => {
-    const newList = replaceItemAtIndex(todoList, index, {
-      ...item,
-      isComplete: !item.isComplete,
-    })
-
-    setTodoList(newList)
-  }
-
-  const deleteItem = () => {
-    const newList = removeItemAtIndex(todoList, index)
-
-    setTodoList(newList)
-  }
-
-  return (
-    <div>
-      <input type="text" value={item.text} onChange={editItemText} />
-      <input
-        type="checkbox"
-        checked={item.isComplete}
-        onChange={toggleItemCompletion}
-      />
-      <button onClick={deleteItem}>X</button>
-    </div>
-  )
-}
-
-const replaceItemAtIndex = (arr, index, newValue) => [
-  ...arr.slice(0, index),
-  newValue,
-  ...arr.slice(index + 1),
-]
-
-const removeItemAtIndex = (arr, index) => [
-  ...arr.slice(0, index),
-  ...arr.slice(index + 1),
-]
 
 export default TodoList
